part2/requirements.txt
datetime
paho-mqtt
